requirements.txt
Django
django-extensions
Werkzeug
PyMySQL
mysqlclient
Pillow
qrcode
openpyxl
gunicorn
psycopg2-binary
pyOpenSSL
